Separate out `android_build.h` from `build_api.h`
Bug: b/422579325
