+++ base/cvd/cuttlefish/host/commands/cvd/fetch/BUILD.bazel
@@ -65,6 +65,7 @@ cc_library(
         "//cuttlefish/host/libs/web/cas:cas_downloader",
         "//cuttlefish/host/libs/web/http_client",
         "//cuttlefish/host/libs/web/http_client:curl_global_init",
+        "//cuttlefish/host/libs/web:android_build",
         "//cuttlefish/host/libs/web:android_build_api",
         "//cuttlefish/host/libs/web:android_build_string",
         "//cuttlefish/host/libs/web:caching_build_api",
