+++ base/cvd/cuttlefish/host/libs/web/BUILD.bazel
@@ -8,6 +8,24 @@ package(
 exports_files([".clang-tidy"])
 
 cc_library(
+    name = "android_build",
+    srcs = ["android_build.cc"],
+    hdrs = ["android_build.h"],
+    copts = COPTS,
+    strip_include_prefix = "//cuttlefish",
+    deps = [
+        "//cuttlefish/common/libs/utils:environment",
+        "//libbase",
+    ],
+)
+
+clang_tidy_test(
+    name = "android_build_clang_tidy",
+    srcs = [":android_build"],
+    tags = ["clang_tidy", "clang-tidy"],
+)
+
+cc_library(
     name = "android_build_api",
     srcs = ["android_build_api.cpp"],
     hdrs = ["android_build_api.h"],
@@ -18,11 +36,12 @@ cc_library(
         "//cuttlefish/common/libs/utils:environment",
         "//cuttlefish/common/libs/utils:files",
         "//cuttlefish/common/libs/utils:result",
+        "//cuttlefish/host/libs/web/cas:cas_downloader",
+        "//cuttlefish/host/libs/web/http_client",
+        "//cuttlefish/host/libs/web:android_build",
         "//cuttlefish/host/libs/web:android_build_string",
         "//cuttlefish/host/libs/web:build_api",
-        "//cuttlefish/host/libs/web/cas:cas_downloader",
         "//cuttlefish/host/libs/web:credential_source",
-        "//cuttlefish/host/libs/web/http_client",
         "//libbase",
         "@jsoncpp",
     ],
@@ -102,6 +121,7 @@ cc_library(
         "//cuttlefish/common/libs/utils:result",
         "//cuttlefish/host/libs/web/cas:cas_downloader",
         "//cuttlefish/host/libs/web/http_client",
+        "//cuttlefish/host/libs/web:android_build",
         "//cuttlefish/host/libs/web:android_build_api",
         "//cuttlefish/host/libs/web:android_build_string",
         "//cuttlefish/host/libs/web:build_api",
